Make reference_workflow a decorator
