+++ flytekit/__init__.py
@@ -8,10 +8,9 @@ from flytekit.annotated.launch_plan import LaunchPlan
 from flytekit.annotated.map_task import maptask
 from flytekit.annotated.reference import get_reference_entity
 from flytekit.annotated.reference_entity import TaskReference, WorkflowReference
-from flytekit.annotated.reference_task import reference_task
 from flytekit.annotated.resources import Resources
-from flytekit.annotated.task import task
-from flytekit.annotated.workflow import WorkflowFailurePolicy, workflow
+from flytekit.annotated.task import reference_task, task
+from flytekit.annotated.workflow import WorkflowFailurePolicy, reference_workflow, workflow
 from flytekit.loggers import logger
 
 __version__ = "0.16.0a2"
