+++ flytekit/annotated/reference.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Dict, Type
 
 from flytekit.annotated.launch_plan import ReferenceLaunchPlan
-from flytekit.annotated.reference_task import ReferenceTask
+from flytekit.annotated.task import ReferenceTask
 from flytekit.annotated.workflow import ReferenceWorkflow
 from flytekit.common.exceptions.user import FlyteValidationException
 from flytekit.models.core import identifier as _identifier_model
